# docs/swagger.yaml
basePath: /api
definitions:
  handler.AddToDraftResponse:
    description: ID заявки-черновика
    properties:
      calculation_id:
        example: 1
        type: integer
    type: object
...
  handler.DraftGroupDeleteRequest:
    description: ID группы для удаления из черновика заявки
    properties:
      group_id:
        example: 1
        type: integer
    required:
    - group_id
    type: object
  handler.DraftGroupResponse:
    description: ID заявки-черновика после операции
    properties:
      calculation_id:
        example: 1
        type: integer
    type: object
  handler.DraftGroupUpdateRequest:
    description: Данные для обновления группы в черновике (индекс CAVI)
    properties:
      cavi_index:
        example: 7.5
        type: number
      group_id:
        example: 1
        type: integer
    required:
    - group_id
    type: object
  handler.ErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      message:
        example: invalid input
        type: string
      status:
        example: fail
        type: string
    type: object
  handler.GroupCreateRequest:
    description: Данные для создания новой группы пациентов
    properties:
      age_group:
        enum:
        - young
        - middle
        - elderly
        example: young
        type: string
      description:
        example: Эластичные сосуды с низким уровнем жесткости
        type: string
      disease_type:
        enum:
        - diabetes
        - hypertension
        example: diabetes
        type: string
      name:
        example: Молодые пациенты
        type: string
    required:
    - age_group
    - name
    type: object
  handler.GroupResponse:
    description: Информация о группе пациентов
    properties:
      age_group:
        example: young
        type: string
      description:
        example: Эластичные сосуды с низким уровнем жесткости
        type: string
      disease_type:
        example: diabetes
        type: string
      id:
        example: 1
        type: integer
      image_url:
        example: http://localhost:9000/cavi-images/diagrams/1.jpg
        type: string
      is_deleted:
        example: false
        type: boolean
      is_selected:
        example: false
        type: boolean
      name:
        example: Молодые пациенты (до 35 лет)
        type: string
    type: object
  handler.GroupUpdateRequest:
    description: Данные для обновления группы пациентов
    properties:
      age_group:
        enum:
        - young
        - middle
        - elderly
        example: middle
        type: string
      description:
        example: Обновлённое описание
        type: string
      disease_type:
        enum:
        - diabetes
        - hypertension
        example: hypertension
        type: string
      is_selected:
        example: true
        type: boolean
      name:
        example: Обновлённое название
        type: string
    type: object
  handler.ImageUploadResponse:
    description: URL загруженного изображения
    properties:
      image_url:
        example: http://localhost:9000/cavi-images/diagrams/1.jpg
        type: string
    type: object
...
  handler.SuccessResponse:
    description: Стандартный успешный ответ
    properties:
      status:
        example: ok
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cavi.local
    name: API Support
  description: |-
    API для расчёта индекса CAVI (Cardio-Ankle Vascular Index).
    Система позволяет управлять группами пациентов, создавать заявки на расчёт,
    добавлять группы в заявки и проводить модерацию.
  title: CAVI Calculator API
  version: "1.0"
paths:
...
  /cavi-calculations/draft/groups:
    delete:
      consumes:
      - application/json
      description: Удаляет группу из заявки-черновика текущего пользователя.
      parameters:
      - description: ID группы
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.DraftGroupDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Группа удалена
          schema:
            $ref: '#/definitions/handler.DraftGroupResponse'
        "400":
          description: Неверный ID или группа не найдена в черновике
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить группу из черновика
      tags:
      - calculation-groups
    put:
      consumes:
      - application/json
      description: Обновляет индекс CAVI для группы в заявке-черновике текущего пользователя.
      parameters:
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.DraftGroupUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Группа обновлена
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Неверные данные или черновик не найден
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить группу в черновике
      tags:
      - calculation-groups
  /cavi-groups:
    get:
      description: Возвращает список всех групп пациентов с возможностью фильтрации
        по названию, возрастной группе и типу заболевания.
      parameters:
      - description: Поиск по названию (частичное совпадение)
        in: query
        name: title
        type: string
      - description: Фильтр по возрастной группе
        enum:
        - young
        - middle
        - elderly
        in: query
        name: age_group
        type: string
      - description: Фильтр по типу заболевания
        enum:
        - diabetes
        - hypertension
        in: query
        name: disease_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список групп
          schema:
            items:
              $ref: '#/definitions/handler.GroupResponse'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Список групп
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Создаёт новую группу пациентов. Доступно только модераторам.
      parameters:
      - description: Данные новой группы
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.GroupCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Группа создана
          schema:
            $ref: '#/definitions/handler.GroupResponse'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Требуется роль модератора
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать группу
      tags:
      - groups
  /cavi-groups/{id}:
    delete:
      description: Мягкое удаление группы (is_deleted = true). Также удаляет изображение.
        Доступно только модераторам.
      parameters:
      - description: ID группы
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Группа удалена
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Неверный ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Требуется роль модератора
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить группу
      tags:
      - groups
    get:
      description: Возвращает детальную информацию о группе пациентов по её ID.
      parameters:
      - description: ID группы
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о группе
          schema:
            $ref: '#/definitions/handler.GroupResponse'
        "400":
          description: Неверный ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Группа не найдена
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить группу
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Обновляет данные группы пациентов. Доступно только модераторам.
      parameters:
      - description: ID группы
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.GroupUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Группа обновлена
          schema:
            $ref: '#/definitions/handler.GroupResponse'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Требуется роль модератора
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить группу
      tags:
      - groups
  /cavi-groups/{id}/add-to-draft:
    post:
      description: Добавляет группу в заявку-черновик текущего пользователя. Если
        черновика нет — создаёт новый.
      parameters:
      - description: ID группы
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Группа добавлена в черновик
          schema:
            $ref: '#/definitions/handler.AddToDraftResponse'
        "400":
          description: Неверный ID или группа уже в черновике
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Добавить в черновик
      tags:
      - groups
  /cavi-groups/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает или заменяет изображение группы. Доступно только модераторам.
      parameters:
      - description: ID группы
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: Файл изображения (JPEG, PNG)
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Изображение загружено
          schema:
            $ref: '#/definitions/handler.ImageUploadResponse'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Требуется роль модератора
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Загрузить изображение
      tags:
      - groups
...
securityDefinitions:
  BearerAuth:
    description: 'JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"